--- src/pages/Admin/Charity/ContributorVerification/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ContributorVerification";
