+++ src/pages/Admin/Charity/OtherSettings/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./OtherSettings";
